Move combine-html files into combine package
Updated package line
Removed combine-html from Makefile
Removed root go mod and go sum
Ran go get and go mod tidy
