+++ app/combine/rewrite.go
@@ -1,4 +1,4 @@
-package main
+package combine
 
 import (
 	"encoding/base64"
