+++ app/combine/rewrite_test.go
@@ -1,4 +1,4 @@
-package main
+package combine
 
 import (
 	"bytes"
