--- combine-html/Makefile
@@ -1,11 +1,0 @@
-GO ?= go
-
-.PHONY: build clean
-
-build: combine-html
-
-combine-html: main.go rewrite.go
-	$(GO) build $(GOBUILDFLAGS) -o $@ $^
-
-clean:
-	rm -rf combine-html
