--- combine-html/go.mod
@@ -1,5 +1,0 @@
-module github.com/cprbucat2/team-feedback/combine-html
-
-go 1.20
-
-require golang.org/x/net v0.11.0
